# Do not check if key is present before removing item in AuthenticationProperties