* Do not check if key is present before removing item, this eliminates double lookup.
* Use `Nullable<T>.GetValueOrDefault()` instead of `Nullable<T>.Value` when it is known to have a value. See #1063 for more info.
* Add some unit tests.

Eliminating the extra lookup yields the following result:

**Before:**

|                  Method |      Mean |     Error |    StdDev |         Op/s |  Gen 0 | Allocated |
|------------------------ |----------:|----------:|----------:|-------------:|-------:|----------:|
|    NullValue_KeyPresent | 123.84 ns | 2.1405 ns | 1.8975 ns |  8,074,763.9 | 0.0036 |     328 B |
| NullValue_KeyNotPresent |  41.63 ns | 0.5963 ns | 0.5577 ns | 24,022,181.2 | 0.0021 |     192 B |

**After:**

|                  Method |      Mean |     Error |    StdDev |         Op/s |  Gen 0 | Allocated |
|------------------------ |----------:|----------:|----------:|-------------:|-------:|----------:|
|    NullValue_KeyPresent | 107.04 ns | 0.9192 ns | 0.8598 ns |  9,342,589.5 | 0.0038 |     328 B |
| NullValue_KeyNotPresent |  39.53 ns | 0.5507 ns | 0.4882 ns | 25,300,109.1 | 0.0021 |     192 B |

Benchmark is available [here](https://gist.github.com/drieseng/96ad8c2680b3ac5b84b6fc2ac0bae100).