@drieseng 2018-11-11
@poke That's knowledge you gain for free by aging :-)