@drieseng 2018-11-14
@Tratcher I just happened to stumble upon it. I didn't come up in profiling session or anything.