@poke 2018-11-14
@Tratcher I would assume that this was motivated by the `Value` to `ValueOrDefault()` change and then the key checks were seen in the code.. 😄